server/requirements.txt
django
djangorestframework
pika
